Issue #199: Installer optimization
Simplified and future-proofed Installer::generateConfigFile.

Old code would break if the default values changed in the sample config file.

New code looks for ['something'] on each line of the sample config file, and if 'something' matches one of the config values we want to write, that line will be re-written into:
...= 'new_value';...