Three small separate rt-related mods.
First, display a '+' after the RT count when the count from twitter has maxed out, to indicate that this has happened and that the count may be higher.
A new non-persistent rt_threshold field is added to post objects as they are built, then the templates check if this field is set and add a '+' if so.
In the post DAO, if have a retweeted post, but its rt value is 0, set to 1.
In the crawler, check that don't have retweeted post id before trying to detect a retweet the hard way.