Issue #779: Add option to disable API.
Set API on by default.
Added missing documentation for replied to and retweeted posts.
Clarified what happens when you overflow into pages of results that do not exist.
Added links to documentation in PostAPIController.
Added regression tests for disabling the API to TestOfPostAPIController.
mwilkie fixed a bug with options not being checked if they default to true.
Fixed a bug in the config for values that had a db_value of false and default
of true always returning the default because of a failing conditional.