Fix working directory & table prefix assumptions in tests
- Make paths to init.tests.php relative to __FILE__ so they don't make
  assumptions about the working directory

- Fix tests that assume the prefix is tu_

- Fix a few places that assume a trailing slash after directories

- Restore the configuration file if it's moved out of the way during
  testing and an exception occurs before it's moved back

- Minor updates to tests/README.md

Remove /tests from THINKUP_BASE_URL

THINKUP_BASE_URL was set to /tests when tests were run from the root
directory.

Make sure there's a slash between root dir and tests/view

Accept "Access Denied" as a correct failure on bad database

When trying to install to a bad database name, the test assumed that
access would be granted, and an "Unknown database" error would result.
If access is denied, accept that as a correct failure as well.

Use loopback for testing bad database host

When using a "helpful" dns resolver that returns an IP address for
unresolvable hosts (see e.g. http://dns-opt-out.comcast.net/), using an
unresolvable hostname like "localcheese" causes the test to hang while
attempting to connect to port 3306 on an IP address that's dropping
packets.

Instead, we try to connect to a loopback address, 127.0.0.2, which
(hopefully) isn't accessible so the connection is refused immediately.

Fix case of MySQL in class names

More fixes for working directory & table prefix assumptions in tests

- Make paths relative to __FILE__ so they don't make assumptions about the
  working directory

- Fix tests that assume the prefix is tu_

- Fix a few places that assume a trailing slash after directories

Fix for non-standard prefix that contains "tu_"

Use #prefix# in query passed to newMigrationFiles