Use loopback for testing bad database host
When using a "helpful" dns resolver that returns an IP address for
unresolvable hosts (see e.g. http://dns-opt-out.comcast.net/), using an
unresolvable hostname like "localcheese" causes the test to hang while
attempting to connect to port 3306 on an IP address that's dropping
packets.

Instead, we try to connect to a loopback address, 127.0.0.2, which
(hopefully) isn't accessible so the connection is refused immediately.