Fix ability to add a Facebook account under https
When running under https, the redirect url generated when creating an
authorization url correctly set the protocol to https if the ThinkUp web
server is running under https, but always set the redirect url protocol
to http when creating the url to request an access token.  This caused
Facebook to reject the request for the access token if the app was
configured with the https url.