Handle SERVER_NAME not being set
When constructing the redirect url to get an access token, use the
HTTP_HOST header value if SERVER_NAME is not defined.  HTTP_HOST should
actually be the default since this is what is used to construct the
authorization url in BaseFacebook::getLoginUrl, but SERVER_NAME is kept
as the default to make this change minimally intrusive; it fixes the
case when SERVER_NAME is not defined by the web server and shouldn't
break any cases that are currently working nor any of the existing
tests.

The HTTP_HOST was also changed in
TestOfFacebookPluginConfigurationController::setUp to match SERVER_NAME
since this is the normal case.  Other cases should be handled by special
tests.
Closes #1049