Copyedit - Developer How to; Consistency in code block formatting
Update codeblock lines to use the :: preceding para format which
also ends up adding the code in more distinct CSS style.

Modify DB page needs a clear heading for step 3 in order to really
delivery the punch of the joke. Otherwise it's lost in step 2.

Rewords a few sentences in order to not have the codeblock break
up the sentence in mid-stream which gets rid of a pause when reading.

Closes #1594
[ci skip]