Plugins link on account management now always takes you to plugins.
Previously, navigating to a specific plugin added the p= parameter
which caused only that plugin to load in the plugins section,
and not the plugin list.  This loads the list and active plugin
and adds javascript to allow opening that section as needed.

Updated Controller Test to expect body and installed_plugins to
both exist.