Handle Facebook business accounts.
In the past, facebook allowed business owners to sign up for an account with
no wall, profile, etc, that was just for managing pages and ads.

They no longer do so.  The /me api for these accounts does not return
a name field, which is sad, because we use that as the key in the instances
table.

For reference, I found these stack overflow on this topic and the suggestions
all come down to "check for a name field":
http://stackoverflow.com/questions/14719211/how-can-i-know-if-a-facebook-account-is-personal-account-or-a-company-type-accou
http://stackoverflow.com/questions/10857138/how-to-obtain-username-name-for-a-company-facebook-account

This simply checks for a name parameter and then shows the user an error.