Improve twitter photo handling.
Using many twitter clients, images end up in the entities->media with the url and the image source,
such as the official twitter client.  These photos are then replaced with a t.co short link
which is not easily/possibly expanded to the actual image url.  But this data is available at
crawl time for us to grab and store properly in the links table.

References #1765