[foursquare] update for new photo data structure
Foursquare changed the photo object structure at some point. Checkins
with photos attached were causing database integrity errors due to the
$photo->url field not existing, causing the crawler to try to insert
NULLs into a NOT NULL field.

Not sure if this is changed for all checkins or just checkins since a
certain date, so preserved use of the old field in an elseif.