Issue #2236 - Fixes Forgotton Passwords for self-hosting
The forgotton passwords controller had a check for the value of the
submit button on the template. However, the redesign changed the value
of the button, and the controller has never been changed to catch up.

This commit makes that change, updates the controller tests and adds an
end-to-end test to catch the regression in future.

Closes #2236