Fixes Dictionary insight test by using generated dates
The previous version of the NewDictionaryTest was scanning for a fixed
date string, which only worked for a month, since the date string is generated
relative to the current date. This adjusts the test to use the relative
date (as in a previous test in the same class), which should allow it to
be date invariant.