Proof-of-Concept docker setup for ThinkUp
The steps required to get the containers up and running can be found in
the `Dockerfile`.

It is purposely self-contained in only two files. If desired, the setup
can and should be expanded to include, among other things, a full
configuration file for the ssmtp tool, even less manual steps (as outlined
in the `Dockerfile`), potentially by providing an out-of-the-box account
that doesn't require a manual browser setup, and automatically run migrations.
The relay host setup could also be improved.

But it's a start.